% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq-shiny}
\alias{faq-shiny}
\alias{faqOutput}
\alias{renderFaq}
\title{Shiny bindings for faq}
\usage{
faqOutput(outputId, width = "100\%", height = "100\%")

renderFaq(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{expr}{An expression that generates a faq}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using faq within Shiny
applications and interactive Rmd documents.
}
