% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq}
\alias{faq}
\title{FAQ page}
\usage{
faq(
  data,
  width = NULL,
  height = NULL,
  elementId = NULL,
  faqtitle = "Frequently Asked Questions"
)
}
\arguments{
\item{data}{df with question and answer columns}

\item{width}{width of this widget}

\item{height}{height of this widget}

\item{elementId}{ellement ID of this widget}

\item{faqtitle}{title for this widgets}
}
\description{
Create FAQ page
}
\examples{
df <- data.frame(
question = c("Question1", "Question2", "Question3"),
answer = c("answer for question1",
           "question2 answer",
           "answer3")
)
faq::faq(data = df, elementId = "faq", faqtitle = "Frequently Asked Questions")

}
